Changed position variables from uint16 to uint8

+++ firmware/src/can_app.c
@@ -5,7 +5,7 @@ volatile can_app_flags_t can_app_flags;
 uint16_t can_app_checks_without_mic19_msg;
 uint32_t can_app_send_state_clk_div;
 uint32_t can_app_send_adc_clk_div;
-volatile uint16_t str_whl_position;
+volatile uint8_t str_whl_position;
 volatile hbridge_flags_t hbridge_flags;
 
 /**
