+++ firmware/src/machine.h
@@ -67,7 +67,7 @@ typedef struct measurements{
     uint16_t    batvoltage_avg;       
     uint16_t    batvoltage_avg_sum_count;
     uint64_t    batvoltage_avg_sum;   
-    uint16_t    position_avg;       
+    uint8_t    position_avg;       
     uint16_t    position_avg_sum_count;
     uint64_t    position_avg_sum;   
     uint16_t    batcurrent_avg;
